[SPACETIME-BACK] include Authentication on memories routes

+++ server/src/routes/memories.ts
@@ -3,21 +3,36 @@ import { prisma } from '../lib/prisma'
 import { z } from 'zod'
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get('/memories', async () => {
+  app.addHook('preHandler', async (req) => {
+    await req.jwtVerify()
+  })
+
+  app.get('/memories', async (req) => {
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: req.user.sub,
+      },
       orderBy: { createdAt: 'asc' },
     })
 
     return memories.map((memory) => {
-      return {
+      const result = {
         id: memory.id,
         coverUrl: memory.coverUrl,
-        except: memory.content.substring(0, 120).concat('...'),
+        except: memory.content,
       }
+
+      if (memory.content.length <= 120) {
+        result.except = memory.content.concat('.')
+      } else {
+        result.except = memory.content.substring(0, 120).concat('...')
+      }
+
+      return result
     })
   })
 
-  app.get('/memories/:id', async (req) => {
+  app.get('/memories/:id', async (req, res) => {
     // const { id } = req.params
 
     const paramsSchema = z.object({
@@ -29,6 +44,12 @@ export async function memoriesRoutes(app: FastifyInstance) {
     const memory = await prisma.memory.findUniqueOrThrow({
       where: { id },
     })
+
+    if (!memory.isPublic && memory.userId !== req.user.sub) {
+      return res.status(401).send({
+        message: 'Essa memoria não é publica ou você não é o dono dela.',
+      })
+    }
 
     return memory
   })
@@ -47,14 +68,14 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: '856baf6c-07c9-4a9f-981c-73690cfbccec',
+        userId: req.user.sub,
       },
     })
 
     return memory
   })
 
-  app.put('/memories/:id', async (req) => {
+  app.put('/memories/:id', async (req, res) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -69,6 +90,16 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
     const { content, coverUrl, isPublic } = bodySchema.parse(req.body)
 
+    const findMemory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if (findMemory.userId !== req.user.sub) {
+      return res.status(401).send({ message: 'Deu ruim ao atualizar' })
+    }
+
     const newMemory = await prisma.memory.update({
       where: { id },
       data: {
@@ -81,12 +112,22 @@ export async function memoriesRoutes(app: FastifyInstance) {
     return newMemory
   })
 
-  app.delete('/memories/:id', async (req) => {
+  app.delete('/memories/:id', async (req, res) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
 
     const { id } = paramsSchema.parse(req.params)
+
+    const findMemory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if (findMemory.userId !== req.user.sub) {
+      return res.status(401).send({ message: 'Deu ruim ao deletar' })
+    }
 
     await prisma.memory.delete({
       where: { id },
